requirements.txt
# === Machine Learning ===
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.3
imbalanced-learn==0.12.4
shap==0.46.0
optuna==4.1.0

# === Visualização ===
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1

# === LLM e Agente ===
langchain==0.3.14
langchain-groq==0.2.4
langchain-community==0.3.14
langgraph==0.2.60

# === API ===
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4

# === Interface ===
streamlit==1.41.1
requests==2.32.3

# === Utilitários ===
joblib==1.4.2
python-dotenv==1.0.1
httpx==0.28.1
